requirements.txt
pandas
numpy
tensorflow
keras
nltk
scikit-learn
python-box
ensure